Change deprecated elements in build file #1
Dependencies section was using 'compile' instead of 'implementation'.

appcompat version was targetting 23.0.3 which I thought might explain why
changing the targetSdkVersion to 23 worked. However, changing to
28.0.0 (and the build tools to 28.0.3) only prevents the Gradle
warnings. The camera is still crashing. Going to play around with this a
bit more before reverting to the workaround of using a lower target SDK.