Rewrites file handling to get working images
Thanks to log messages and breakpoints, it became evident that most of
onActivityResult was not being executed. Rewriting the file creation
process got around that, while rewriting the image handling code based
on the Photo Basics section of the Android developer site allows the
image to appear in the email. Closes #1