Remove hard-coded colours from vector images
Each vector image is composed of discrete objects that can be controlled
separately. This commit removes the hard-coded colours used for each
vector object and ties them to the colours defined in colors.xml.

This means that changing, say, the primary_color value will update not
just the app colours, but also those of each image and icon too.