Adds a simple menu item for ordering
I felt that having the 'Send Order' button inside the order tab might be
a bit hidden, so I've put an app-wide menu item instead. For example, if
someone selects a t-shirt and goes straight to the 'Collection' tab,
they could miss how to place an order.

It's currently not functional, will fix that in a future update.