Recreates email summary text in a separate class
Currently, responsibity for recreating the email summary lies with the
FragmentOrders class. However, the app is moving in a direction where it
should probably belong to the main activity instead, as it will be
drawing from values shared among different fragments using shared
preferences.

This commit makes an attempt to begin this process and demonstrates its
progress by outputting the default values to a Toast message, but will
eventually replace that with opening the email app.