Fixes "Can't attach empty file" message
The Orders fragment now provides a more robust indication of whether a
photo has been taken or not. Previously, opening a camera intent and
cancelling it left some variables initialised, which gave false readings
when determining whether or not to attach a photo to the email. This has
now been fixed. #11