Attempts to save collection state
Not working yet: trying to store the state of the collection fragment so
it can be restored later. It seems to be half-remembering the previous
selection (i.e. although nothing appears to be selected, it still knows
to 'deselect' the same item when pressed again). Could be just a visual
glitch, but needs further investigation. #13