Adds order details to SharedPreferences
Unfortunately, they're not being picked up correctly. The
SharedPreferences are being set in onSaveInstanceState, but I need to
check if this gets triggered when the order button is clicked (probably
not if the order fragment is already open). There is also an issue where
the address is ignored even if the collection address is empty. Possibly
a related issue.